Rename NoteSkin namespace to NoteSkins
