+++ OpenChart/src/Noteskins/NoteSkinManager.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text.RegularExpressions;
 
-namespace OpenChart.NoteSkin
+namespace OpenChart.NoteSkins
 {
     ///
     /// <summary>
